search routine
I have created this stored procedure on the MySQL database.  It will accept up to five parameters with values as long as the parameters are the exact column names in the tables.  It will return the list of employee names and ids.

call this using this format:

CALL search(param1, value1, param2, value2, param3, value3, param4, value4, param5, value5)